+++ extensions/readwise-to-tana/src/index.tsx
@@ -14,8 +14,9 @@ import { useBooks } from './useApi'
 
 export default function Command() {
   const [template, setTemplate] = React.useState<string>('')
+  const [category, setCategory] = React.useState<string>('all')
 
-  const { data, isLoading } = useBooks()
+  const { data, isLoading } = useBooks({ category })
   const { pop } = useNavigation()
 
   React.useEffect(() => {
@@ -45,6 +46,9 @@ export default function Command() {
       highlightLocation,
       id,
       source,
+      title,
+      readwiseUrl,
+      url,
     } = values
     let t = '%%tana%%'
 
@@ -57,6 +61,11 @@ export default function Command() {
     t += category ? `\n  - ${category}:: {{category}}` : ''
     t += source ? `\n  - ${source}:: {{source}}` : ''
     t += coverImageUrl ? `\n  - ${coverImageUrl}:: {{cover_image_url}}` : ''
+    t += readwiseUrl
+      ? `{{#if highlights_url}}\n  - ${readwiseUrl}:: {{highlights_url}}{{/if}}`
+      : ''
+    t += url ? `{{#if source_url}}\n  - ${url}:: {{source_url}}{{/if}}` : ''
+    t += title ? `\n  - ${title}:: {{title}}` : ''
 
     let highlights = '\n\n{{#each highlights}}'
 
@@ -81,9 +90,34 @@ export default function Command() {
   }
 
   return (
-    <List isLoading={isLoading} searchBarPlaceholder="Filter Books">
+    <List
+      searchBarAccessory={
+        <List.Dropdown
+          tooltip="Select Category"
+          onChange={setCategory}
+          value={category}
+        >
+          <List.Dropdown.Item value="all" title="All Categories" />
+          {['articles', 'books', 'podcasts', 'supplementals', 'tweets'].map(
+            (value) => (
+              <List.Dropdown.Item
+                key={value}
+                value={value}
+                title={value.charAt(0).toUpperCase() + value.substring(1)}
+              />
+            )
+          )}
+        </List.Dropdown>
+      }
+      isLoading={isLoading}
+      searchBarPlaceholder="Filter Library"
+    >
       {data?.results.length === 0 ? (
-        <List.EmptyView title="No books found" />
+        <List.EmptyView
+          title={
+            category === 'all' ? 'No results found' : `No ${category} found`
+          }
+        />
       ) : (
         data?.results.map((book) => (
           <List.Item