+++ extensions/readwise-to-tana/src/useApi.ts
@@ -7,9 +7,11 @@ export type Book = {
   author: string
   category: string
   cover_image_url: string
+  highlights_url: string
   id: string
   num_highlights: number
   source: string
+  source_url: string | null
   title: string
 }
 
@@ -39,8 +41,14 @@ const getHeaders = () => {
   }
 }
 
-export const useBooks = () => {
-  return useFetch<{ results: Book[] }>(`${baseUrl}/books/`, getHeaders())
+export const useBooks = ({ category }: { category: string }) => {
+  const url = new URL(`${baseUrl}/books/`)
+
+  if (category !== 'all') {
+    url.searchParams.set('category', category)
+  }
+
+  return useFetch<{ results: Book[] }>(url.toString(), getHeaders())
 }
 
 export const useBook = (id: string) => {